# Use only level 1 and 2 with python-isal